+++ src/components/project-card/index.tsx
@@ -52,7 +52,8 @@ export const ProjectCard: React.FC<IProject> = (props) => {
 
   return (
     <Card
-      direction={{ base: "column", sm: "row" }}
+      mx="4"
+      direction={{ base: "column", lg: "row" }}
       overflow="hidden"
       variant="filled"
       my={8}
@@ -62,7 +63,7 @@ export const ProjectCard: React.FC<IProject> = (props) => {
         background={is_private ? COLORS.warning500 : COLORS.primaryColor}
         color={"white"}
         position={"absolute"}
-        top={"50%"}
+        top={{ base: 8, lg: "50%" }}
         right={4}
       >
         <TagLabel
@@ -75,13 +76,12 @@ export const ProjectCard: React.FC<IProject> = (props) => {
       </Tag>
       <Image
         objectFit="cover"
-        // maxW={{ base: "100%", sm: "200px" }}
         src="https://images.unsplash.com/photo-1667489022797-ab608913feeb?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxlZGl0b3JpYWwtZmVlZHw5fHx8ZW58MHx8fHw%3D&auto=format&fit=crop&w=800&q=60"
         alt="Caffe Latte"
         pt={4}
         pb={4}
         pl={4}
-        htmlWidth="240px"
+        width={{ base: "100%", lg: "30%" }}
         borderRadius="3xl"
       />
       <CardBody>
@@ -135,6 +135,7 @@ export const ProjectCard: React.FC<IProject> = (props) => {
               size="lg"
               colorScheme={getProjectStatusColor(status)}
               width={"fit-content"}
+              mr={2}
             >
               <TagLabel>{status}</TagLabel>
             </Tag>
