+++ src/pages/landing/index.ts
@@ -1,0 +1,1 @@
+export { Landing } from "./landing";
